Refactored out drop area
