+++ nomenclate/core/env.yml
@@ -207,11 +207,11 @@ options:
     caudal: [cau, caud]
 
   compoutput:
-    LSC: 'lighting slap comp'
-    IG: 'input graded'
-    I: 'input'
-    ASC: 'asset slap comp'
-    R: 'roto'
+    lighting_slap_comp: 'LSC'
+    input_graded: 'IG'
+    input: 'I'
+    asset_slap_comp: 'ASC'
+    roto: 'R'
 
   status:
     preproduction: PREPROD
